sfdp: modify tenant_index type from u16 to u32
* 'Tenant_index' was previously set on 16 bits. This
PR modify this so that it is set on 32 bits and allow
for more tenants to be specified.

* Fix sections which relied on tenant_index being set on
16 bits e.g. invalid tenant_index value was previously considered
as ((u16) ~0) in certain sfdp services, and has been updated to
be (u32) (~0).

Type: improvement

+++ src/plugins/sfdp_services/acl/acl_sample.c
@@ -5,7 +5,7 @@
 #include <acl/acl_sample.h>
 
 static void
-acl_sample_validate_tenant_lc (u16 tenant_idx)
+acl_sample_validate_tenant_lc (u32 tenant_idx)
 {
   sfdp_acl_main_t *vam = &sfdp_acl_main;
   sfdp_main_t *sfdp = &sfdp_main;
@@ -27,7 +27,7 @@ sfdp_acl_sample_tenant_set_acl (sfdp_acl_main_t *vam, u64 tenant_id,
   sfdp_main_t *sfdp = &sfdp_main;
   clib_error_t *err = 0;
   clib_bihash_kv_8_8_t kv = { .key = tenant_id, .value = 0 };
-  u16 tenant_idx;
+  u32 tenant_idx;
   u32 lc;
   u32 *acl_vec = 0;
   if (clib_bihash_search_inline_8_8 (&sfdp->tenant_idx_by_id, &kv))
