+++ src/plugins/sfdp_services/base/interface_input/interface_input.c
@@ -13,7 +13,7 @@ sfdp_interface_input_set_tenant (sfdp_interface_input_main_t *vim,
   sfdp_main_t *sfdp = &sfdp_main;
   clib_bihash_kv_8_8_t kv = { .key = tenant_id, .value = 0 };
   vnet_main_t *vnm = vnet_get_main ();
-  u16 *config;
+  u32 *config;
 
   if (clib_bihash_search_inline_8_8 (&sfdp->tenant_idx_by_id, &kv))
     return clib_error_return (0, "Tenant with id %d not found");
@@ -21,12 +21,12 @@ sfdp_interface_input_set_tenant (sfdp_interface_input_main_t *vim,
   vec_validate (vim->tenant_idx_by_sw_if_idx, sw_if_index);
   config = vim->tenant_idx_by_sw_if_idx + sw_if_index;
 
-  if (config[0] == ((u16) ~0) && unset)
+  if (config[0] == ((u32) ~0) && unset)
     return clib_error_return (
       0, "Outside tenant %d is not configured on interface %U", tenant_id,
       format_vnet_sw_if_index_name, vnm, sw_if_index);
 
-  if (config[0] != (u16) (~0) && !unset)
+  if (config[0] != (u32) (~0) && !unset)
     return clib_error_return (0, "Interface %U is already configured",
 			      format_vnet_sw_if_index_name, vnm, sw_if_index);
 
@@ -41,7 +41,7 @@ sfdp_interface_input_set_tenant (sfdp_interface_input_main_t *vim,
     {
       vnet_feature_enable_disable ("ip4-unicast", "sfdp-interface-input",
 				   sw_if_index, 0, 0, 0);
-      config[0] = (u16) (~0);
+      config[0] = (u32) (~0);
     }
 
   return 0;
@@ -57,7 +57,7 @@ sfdp_interface_input_add_del_sw_interface (vnet_main_t *vnm, u32 sw_if_index,
     {
       vec_validate (vim->tenant_idx_by_sw_if_idx, sw_if_index);
       for (int i = old_size; i <= sw_if_index; i++)
-	vim->tenant_idx_by_sw_if_idx[i] = (u16) (~0);
+	vim->tenant_idx_by_sw_if_idx[i] = (u32) (~0);
     }
   return 0;
 }
